django-app/requirements.txt
asgiref==3.2.3
Django==3.0.2
django-annoying==0.10.6
django-filter==2.2.0
django-js-reverse==0.9.1
django-model-utils==4.0.0
django-rest==0.0.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
factory-boy==2.12.0
Faker==3.0.0
Markdown==3.1.1
names==0.3.0
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
text-unidecode==1.3
